Adds a flag to determine whether to reload the rules in policy
Initing a Enforcer class with rules can not set rules,  because the policy
file is modified , the load_rules always try to load the rules from the cache
or config file when checks the policy. This patch adds a flag to determine
whether to reload the rules from policy file.

Closes-Bug: #1279198

Change-Id: Ife84189be4b86a3ee90da4539ff2dbed125be23d