Fix tests
Test files from tests/var weren't included in the new repo so the
tests would fail with an exception due to missing files. To fix this
the files are copied from oslo-incubator.

Also, Python 3 requires absolute imports so this was fixed in
oslo_policy.policy.

The imports were out of order so the pep8 tests failed.

Change-Id: I1ee6c2aed9565e5472306bc715ac42b74de6c39f