Do not log on missing or empty policy_dirs
The "missing directory" message for the default policy-directory is
being hit constantly and requiring work-arounds
(e.g. I335a661d307a330e8b59883ef60c787c6590eb42)

It was turned down from a warning in
I183597f23b3a5eb44debd73f7a4a4444ca0d22fd.  As it says there, the
directory missing is considered a normal condition.

Previous revisions of this change have modified this log to only fire
if policy_dirs is set to something other than the default and to
deprecate the default value of policy_dirs so the directory is not
searched.

The policy_dirs option as written (and exposed, hence limiting change)
is analogous to a PATH setting; we attempt to search each entry but a
missing or empty directory does not indicate a problem.

In a theoretical sense this actually seems quite sound; some setup
might have some specific host policy files in /etc/policy.d and global
ones in /usr/share/policy.d; the non-existence of /etc/policy.d would
not be an error but just mean that on this host only the
/usr/share/policy.d ones are wanted.

However, the existing debug level message about a file being should be
expanded to show the actual file that was loaded.  This allows
investigators insight into what policies have actually been loaded and
from where.

Debugging test-cases are updated to verify that we get debug-level
messages about the loaded files.  Additionally, the documentation for
policy_dirs is expanded to make it clear the directories may be
missing or empty.

Related-Bug: #1421863
Change-Id: Ib749b2a019c89bec811d3da5a369ffcdd27bc76f