Clean up configuration option management
Move the option definitions into the opts module with the other
functions related to options.

Add set_defaults() to override the default settings for options and make
it part of the public API of the library by including it in the
documentation.

Add opts._register() so the options can be registered consistently from
multiple places.

Change-Id: Ib74beac58a07bf5ee0f92c2b7eda8281ef7d7c33