Add Rules.from_dict classmethod
Many services currently construct a Rules instance from a dictionary in
their tests. It would be beneficial to them if they didn't have to
import oslo_policy._parser in order to parse a rule before passing it to
Rules.__init__. Adding a Rules.from_dict classmethod should satisfy the
need to create a Rules instance from a dictionary in tests.

Change-Id: I3d73059277c45d4852f842474c6f1d61daa349b8