deprecate policy_dirs option
It is not clear why this is a configurable option. The original
blueprint just wanted the ability to specify a policy.d infrastructure
to make complex policies easier to enforce. There was no requirement
on it being configurable.

Being configurable just exposes another way in which incompatibilities
in OpenStack projects might be able to emerge. As this merged late in
the incubator lifecycle for the policy code, there are likely few/no
existing users of the option so it should be safe to deprecate it
for the library.  That way we won't break anyone currently using it,
and don't have to commit to supporting it going forward.

Co-Authored-By: Steve Martinelli <stevemar@ca.ibm.com>
Change-Id: I40bcbfbd4ab626fa9bfa050792e9ed3cfb587225