Correct invalid doc references
Since the checks are defined in a private module, the docs for them
weren't being generated. The docs are moved to the public symbols so
that they'll be generated.

Also, there were several references from public docs to private
symbols. These are changed since any references to private symbols
aren't going to be able to link to the private symbol docs since
they're not generated.

Change-Id: I27c666479ecb978ad6bf21dc2ad80ef0265f62c5