Perform basic checks on policy definitions
It is easy for someone to configure a policy to reference a rule
definition that does not exist. In this case the check will fallback to
the default rule(if it exists) and the deployer will never realize
they've incorrectly defined their policy.

It is also possible to configure a set of rule checks with a cycle. This
will fail at runtime with a "RunTimeError: maximum recursion depth
exceeded" and a potential 500 HTTP response from whatever API is in use.

This change adds a check for those cases which will log a warning when
those occur. The check can also be called when a service using
oslo.policy starts up and fail to start for incorrect policy
definitions. The default behavior is just to log the condition in case a
policy file is changed after startup and reloaded. That should not raise
an exception which may go unhandled and kill the service.

Change-Id: If2d1a5df05091e013293ec350861d7ada1753a47