Add additional param to policy.RuleDefault
This patch adds the needed subclass to define policy objects
that now require the description and operations parameters.

The operations parameter when defining a policy rule should follow
the below format.
    - operations : list of dicts containing the api url and verb.
        Ex : operations=[{path='/some/url/', method='GET'}]

This change make sure operators have all the information needed
in a centralized location in order to make decisions and understand
the purpose of a certain policy.

Change-Id: Ie9b335420394166bb39c43e3d26fcc9237ffd1a0