generator: Reimplement wrapping of 'description'
The current implementation of wrapping for the Policy.description field
leaves a lot to be desired. It takes each line in the original
description independently and wraps that, ignoring the fact that the
line may be part of a paragraph, or could be a literal that shouldn't be
wrapped. As an example, imagine that wrapping occurred at 40
characters instead of 70. In this case, the below:

    A sample lines with more than forty characters
    which continues down onto the next line.

would become:

    A sample lines with more than forty
    characters
    which continues down onto the next line.

when clearly, what we want is something like this:

    A sample lines with more than forty
    characters which continues down onto the
    next line.

This is resolved.

In addition, it should be possible to include basic literal blocks and
those should be preserved. For example:

    Here's a sample literal block

        This is a really long literal block but it should not wrap

should be output in the exact same way. This is also resolved.

Note that we're not accounting for things like bullet points. Doing so
would bring us close to rST parser implementation territory, and we
don't want to go there. We can revisit this if someone turns out to want
this feature.

Change-Id: I3ea2aac73e3c0a4f77f3f4097540de01264cd618
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>