fix formatting for empty defaults
When the default for a policy rule is empty, the output is ````. The
rst parser interprets that as a title overline or underline, and
sphinx dies. When we have no default rule emit a string saying so to
avoid emitting an empty literal inline block.

Closes-Bug: 1707246
Change-Id: I774b2de5ff59874dfa67811c094735dd74c8083e
Signed-off-by: Doug Hellmann <doug@doughellmann.com>
(cherry picked from commit 2b99c39c30df432fe79cbfe3db0f21039e324fb4)