External Policy hook should support SSL
We have a HttpCheck that can check rules by calling a
remote server. However currently it just supports "http://"
and not "https://". In this patch, we add a HttpsCheck so
we can support "https://" urls as well. In addition to
ensure security, we add options to:
* Specify a ca cert to validate the server
* Specify cert/key to allow server to validate the client

Another quirk was that currently the payload is being
sent using 'application/x-www-form-urlencoded' as the
content-type. We add another option to send the payload
as 'application/json'. Add this same support to the
HttpCheck as well.

Change-Id: Ic9c5249555fb45010a92432a504c84ad3fbb9ea1