Modification to add additional information in the HTTPCheck request.
make it easier to reuse the invocation logic for check objects

Provide a new private function in oslo_policy._checks to evaluate a
check object. This function protects against API changes to the check
classes by inspecting the set of arguments accepted.

Update Enforcer to use the new function instead of invoking checks
directly.

Update the nested check classes (and, or, not) to use the new function
instead of invoking their sub-rules directly.

Update the way mocks were being used in some tests to replace them
with real minimal classes that implement the necessary APIs.

Simplify a few tests that were confirming multiple behaviors (for
example, the result of a compound check as well as the arguments
passed to its nested rules).

Ensure that we have test cases for invoking nested rules that do and
do not accept the new current_rule argument.

Change-Id: Ib9edd7954d0b977950be536fa9434243b0de7fcf
Signed-off-by: Doug Hellmann <doug@doughellmann.com>