Render deprecated policy names when generating files
We found out the when policy files are generated they don't have
deprecated policies. This can be problematic for services that use
rendered policy files because they could still be using the
deprecated policy name and not the new policy name.

This commit makes sure we render the deprecated policy name
as an alias to the new policy name. This makes it so that operators
don't have to change two policies if they override a deprecated
policy.

Closes-Bug: 1742569

Change-Id: Iaf0c89a035775770ceaa230c65ce8eb195b9d82b