Pass dictionary as creds in policy tests
According to the documentation strings for the ``enforce`` method, the
``creds`` parameter is expected to be a dictionary. The implementation
also treats it like a dictionary.

This commit updates three of the tests in test_policy.py to pass a
dictionary into enforce instead of just a string called 'creds'.

Change-Id: Id05a41b84123aeb57b193bba97cb9e8442587a51