Fix requirements and convert to stestr
This commit fixes two issues that are currently blocking the gate.

The first is that it bumps the Sphinx requirement to be within
acceptable constraints. The second is that it converts oslo.policy to
use stestr instead of testr. This is all being done in one patch
because proposing them individually causes deadlock (the patch to
bump the sphinx requirement fails because we're still using testr and
the patch to convert to stestr fails the requirements-check job).

The following explains the reasoning behind the stestr change.

With the upgrade to oslotest 3.6.0 [0], testr no longer works [1].
This is because oslotest no longer requires testr and we don't depend
on it directly in oslo.policy.

[0] https://github.com/openstack/requirements/commit/d5a3c58f7195517a6083032e41b702c2a0aca431
[1] https://github.com/openstack/oslotest/commit/897823fbd61ae9ba015af38e9fed5a5728de64c2

Change-Id: I6dac4c8e7b39c9b80cc8f3728763e8d783c9e940