Add ability to pass in target data for the oslopolicy-checker
This allows us to test the policy for other services which might have
different or unusual target data formats (such as Barbican). It would be
possible to pass it as a nested dictionary, e.g.:

{
    "target": {
        "secret": {
            "project_id": "my project id"
        }
    }
}

or as a key pair (as oslo.policy would expect):
{
    "target.secret.project_id": "my project id"
}

Both will work (note that this logic was taken from barbican).

This fixes around the limitation that the target is hardcoded to be
"project_id", and thus allows to test more scenarios (such as the
project ID not matching).

Change-Id: Ia9f7462072a8cb142251c8bb5ef19d9a25a98119