Make upgrades more robust with policy overrides
Previously, we'd notify operators of changing or deprecated policies
by logging a warning while loading rules. However, this doesn't
prevent unintended access if an operator is overriding a policy
by its old policy name. In this case, let's make sure we check if the
old policy is being overridden and use that override for the new
policy's check value.

This commit introduces this change along with a few tests. It also
refactors the deprecated rule logic in load_rules() to separate
methods so that it's a little easier to understand where that logic
happens within the load_rules() method without cluttering it.

Co-Authored-By: Juan Antonio Osorio Robles <jaosorior@redhat.com>

Closes-Bug: 1800259
Change-Id: Ice27cdb44241da94693625776037ea6164bbb913