Fully log RBAC enforcement data
Data passed to the RBAC enforce function consistes of 3 items:

* rule name or rule object
* credential data
* target data

Both the credential and target are dicts. When policy enforcement
does not work as expected it's essential to capture the input to
the enforcement engine as to ascertain why the rule did not work
as expected. It would also be highly advantageous if the logging
were in a format that could be digested by other tools
(e.g. oslopolicy-checker).

This patch does two things:

1) It logs the policy relevant input to Enforcer.enforce()

2) It eschews the use of Python's string formatting which may not
fully dump the contents of the dicts and is not easily parsed in favor
of using JSON format which does fully capture the object's content and
can be used in data exchange (and can be read by oslopolicy-checker).

Contents of the credentials dict are filtered to scrub security
sensitive data.

Closes-Bug: #1804073

Change-Id: I4642c57990b145c0e691140970574412682e66a5
Signed-off-by: John Dennis <jdennis@redhat.com>