Fix usage of token fixture in shell tests
The way we were using the token fixture in the shell tests was modifying
the structure. If the test would get run by the same process, it would
then use the modified structure and the test would fail.

This uses deepcopy instead, so this way, we never modify that fixture.

Change-Id: Ib88feee7d7fe72c66b4e8af510f9f28411ac47df