Prevent sensitive target data from being logged
A previous commit made some changes to allow for more robust logging
of RBAC enforcement data:

  I4642c57990b145c0e691140970574412682e66a5

This also included logging of the target data, which is provided by
the service calling policy enforcement.

This commit makes it so that target data is protected from exposing
sensitive information. A good example is doing operations on users
in keystone since keystone would populate the target dictionary
with user information, and possibly passwords.

This issue was found in keystone unit testing while trying to consume
oslo.policy 1.43.0.

Change-Id: I2702df8f3d7c040312eb863f7772b129e0e2c45c