Provide more specific error when namespace is missing
Previously if a non-existent namespace was specified, we just got
a generic KeyError from stevedore that didn't say a whole lot about
what went wrong. You pretty much had to go read the code to figure
out what happened.

This change adds an explicit check for a missing namespace and raises
a KeyError with a more specific error message that explains what is
wrong.

Change-Id: Ia56d4655d70cee78661567188a977f67b7c3ee78
Closes-Bug: 1817953