Only alias when policy names change
Previously, oslo.policy would generate policy files with aliased names
in the event the name was changing for backwards compatibility. This
isn't needed if the name isn't changing and only the check string is
changing.

This patch adds a conditional to the generator logic that only aliases
the old name to the new name if the name is changing. Otherwise, it
only outputs comments about the deprecation.

Co-Authored-By: Ben Nemec <bnemec@redhat.com>
Closes-Bug: 1845209
Change-Id: I89ff60354e4751a5096832023441d2e6166db92a
(cherry picked from commit 9641e5ca0d758376fb7f7f5156056d5206d71046)