Initialize global config object in cli tools
Currently, passing --config-file to a tool like oslopolicy-list-redundant
is ineffective because the projects pass an empty cli arg list to the
conf object when they initialize it. By registering our cli args on the
global conf object, the projects can safely parse cli args in their
call to the conf object so things like --config-file won't be ignored.
This didn't work before because oslo.policy recognizes cli args like
--namespace that aren't recognized by the consuming projects.

This will require followup changes in each project to stop passing an
empty cli arg list to the conf object initialization.  In the meantime,
everything should continue to work as it did before.

Change-Id: Iacd257fc6c351582de45476768e3fd1775317d3c
Closes-Bug: 1849518