Don't parse cli args on the global object in sphinxpolicygen
sphinxpolicygen is calling the generate_sample cli entrypoint when
we aren't actually the command being run. This can cause problems
if the consuming project has cli args that get registered on import
of their modules because we may have parsed args before those modules
get imported. This results in an exception because oslo.config won't
allow cli args to be registered after they've been parsed once.

This change makes use of the existing parameter to generate_sample
that allows us to pass in a local config object on which to register
the cli args. This way we can parse them without affecting the
global config object.

This was the only place I could find that we were doing something
like this so I believe it should eliminate the problem.

Change-Id: I8e9f28b0a15d1ed092d72b983be74fe281708fbe