Deprecate the JSON support for policy_file
JSON support for policy_file has been problematic
since projects started policy-in-code. For example,
generating a sample policy file in JSON results in
all the policy-in-code rules being overridden because
it is not possible to comment out the default rules in JSON.

Asd part of migration of JSON format to YAML, this commit
deprecates the:
1. Deprecate JSON support in oslo.policy.
2. Deprecate JSON output in policy CLI tools including '--format'
option.

Partial implement blueprint policy-json-to-yaml

Change-Id: I5432a8cf80903620f48936cbbfb92ea6b6ff30fa