Fix oslopolicy-j2y-convertor tool for RuleDefault
Description and operators are mandatory parameter for
DocumentedRuleDefault but few services still using the
RuleDefault for exmaple glance
- https://github.com/openstack/glance/blob/1344c45772c1a77107f621632642cf1488e431db/glance/policies/image.py#L16

To make work on oslopolicy-j2y-convertor tool to covert
the JSON file for such services we need to pass description
in preparing DocumentedRuleDefault for this tool to generate
the yaml file with rule description. For such cases, it add
rule name as description.

Also for 'operations', do not write it in generate file
for RuleDefault as that end up with blank space.

Change-Id: I910291a152402051b1eac96c3ec16c3f0bb8bbb7